十觞亦不醉，感子故意长 基本完成坐标变换

+++ src/RunNode.h
@@ -14,7 +14,7 @@ class RunNode : public std::enable_shared_from_this<RunNode>
 private:
     static std::vector<std::shared_ptr<RunNode>> root_;    //所有需要绘制的内容都存储在这个静态向量中
     static uint64_t prev_present_ticks_;
-    static int refresh_interval_;
+    static double refresh_interval_;
 
 private:
     bool is_private_ = false;
